+++ src/configs/types.d.ts
@@ -3,7 +3,7 @@ import type { SupportedSvgoVersions } from "../svgo";
 interface Config {
   readonly ignoreGlob: string;
   readonly isDryRun: boolean;
-  readonly svgoOptionsPath: string;
+  readonly svgoConfigPath: string;
   readonly svgoVersion: SupportedSvgoVersions;
 }