+++ src/filters/__mocks__/pushed-files.ts
@@ -1,0 +1,5 @@
+const NewPushedFilesFilter = jest.fn()
+  .mockReturnValue([() => false, null])
+  .mockName("NewPushedFilesFilter");
+
+export default NewPushedFilesFilter;