+++ test/__mocks__/filters.mock.ts
@@ -1,0 +1,17 @@
+const NewPrFilesFilter = jest.fn()
+  .mockReturnValue([() => false, null])
+  .mockName("NewPrFilesFilter");
+
+const NewPushedFilesFilter = jest.fn()
+  .mockReturnValue([() => false, null])
+  .mockName("NewPushedFilesFilter");
+
+const NewSvgsFilter = jest.fn()
+  .mockReturnValue(() => false)
+  .mockName("NewSvgsFilter");
+
+export default {
+  NewPrFilesFilter,
+  NewPushedFilesFilter,
+  NewSvgsFilter,
+};