+++ test/integration/clients.test.ts
@@ -78,7 +78,7 @@ describe("package clients", () => {
         test.each([
           [[/* no files */]],
           [[{ filename: "foobar", status: "modified" }]],
-        ])("request success (files: `%p`)", async (files) => {
+        ])("request success (files: `%j`)", async (files) => {
           const owner = "pikachu";
           const repo = "pokédex";
           const ref = "commit-1";
@@ -126,15 +126,19 @@ describe("package clients", () => {
           octokit.rest.pulls.listFiles.mockReset();
         });
 
-        test("request success", async () => {
-          const files = [];
+        test.each([
+          [[/* no files */]],
+          [[{ filename: "foobar", status: "modified" }]],
+        ])("request success (files: `%j`)", async (files) => {
           const owner = "pikachu";
           const repo = "pokédex";
           const pullNumber = 42;
           const perPage = 10;
           const page = 1;
 
-          octokit.rest.pulls.listFiles.mockResolvedValueOnce(files);
+          octokit.rest.pulls.listFiles.mockResolvedValueOnce({
+            data: files,
+          });
 
           const [result, err] = await client.pulls.listFiles({
             owner,